moved data related classes into data package
