+++ src/server/model/data/Hasher.java
@@ -1,4 +1,4 @@
-package server.model.loginandregister;
+package server.model.data;
 
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
