+++ src/server/model/data/XMLLogin.java
@@ -1,4 +1,4 @@
-package server.model.loginandregister;
+package server.model.data;
 
 import java.util.Collections;
 import java.util.HashMap;
